fix growth rate calculation, use net profit
we may consider to work out a dynamic p/e that takes normal distribution
into account, since a growing company may varies on p/e.
of course, it is bubble, but can be take it as an reference although
it may not stands for value, at least it stands for some kind of bubble.